requirements.txt
requests
socketwrench
PyYAML